bugfix/IT2425T07-10 backend structuur aangepast en beveiliging tegen dubbele bomen bij uploaden toegevoegd
