+++ backend/models/__init__.py
@@ -1,0 +1,3 @@
+from .user_model import User
+from .tree_model import Tree
+from .revoked_token_model import RevokedToken
