+++ backend/routers/user_router.py
@@ -1,0 +1,52 @@
+from fastapi import APIRouter, Depends, HTTPException, status
+from sqlalchemy.orm import Session
+from database import SessionLocal, get_db
+from services.user_service import get_user_by_username, authenticate_user, create_user
+from services.token_service import create_access_token
+from pydantic import BaseModel
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from datetime import datetime, timedelta, timezone
+import os
+
+router = APIRouter()
+
+class UserCreate(BaseModel):
+    username: str
+    password: str
+
+
+SECRET_KEY = os.getenv("SECRET_KEY")
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+def standard_response(success: bool, message: str, data: dict | None = None):
+    return {
+        "success": success,
+        "message": message,
+        "data": data
+    }
+
+@router.post("/register")
+def register_user(user: UserCreate, db: Session = Depends(get_db)):
+    db_user = get_user_by_username(db, username=user.username)
+    if db_user:
+        raise HTTPException(
+            status_code=400, 
+            detail=standard_response(False, "Username already registered")
+            )
+    return create_user(db, user.username, user.password)
+
+@router.post("/login")
+def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
+    user = authenticate_user(form_data.username, form_data.password, db)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=standard_response(False, "Incorrect username or password"),
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.username}, expires_delta=access_token_expires
+    )
+    return standard_response(True, "Login successful", {"access_token": access_token, "token_type": "bearer"})
